+++ cellconstructor/ThermalConductivity.py
@@ -9,7 +9,6 @@ import os, sys
 import scipy, scipy.optimize
 from scipy import integrate
 import h5py
-import psutil
 
 # Import the Fortran Code
 import symph
@@ -29,6 +28,12 @@ import warnings
 
 import time
 
+__SEEKPATH=False
+try:
+    import phonopy
+    __SEEKPATH=True
+except:
+    __SEEKPATH=False
 
 try:
     from mpi4py import MPI
@@ -308,7 +313,7 @@ def get_kpoints_in_path(path, nkpts, kprim):
     return kpoints, distance, segments
 
 
-def stupid_centering_fc3_v3(tensor3, check_for_symmetries = True, Far = 1):
+def centering_fc3(tensor3, check_for_symmetries = True, Far = 1):
     rprim = tensor3.unitcell_structure.unit_cell.copy()
     irprim = np.linalg.inv(rprim)
     rsup = tensor3.supercell_structure.unit_cell.copy()
@@ -329,7 +334,6 @@ def stupid_centering_fc3_v3(tensor3, check_for_symmetries = True, Far = 1):
         print('ForceTensor most likely not previously centered! ')
         if(check_for_symmetries):
             permutation = thermal_conductivity.third_order_cond_centering.check_permutation_symmetry(tensor3.tensor, tensor3.r_vector2.T, tensor3.r_vector3.T, tensor3.n_R, natom)
-            print(psutil.virtual_memory().percent)
             if(not permutation):
                 print('Permutation symmetry not satisfied. Forcing symmetry! ')
                 fc3 = thermal_conductivity.third_order_cond_centering.apply_permutation_symmetry(tensor3.tensor, tensor3.r_vector2.T, tensor3.r_vector3.T, tensor3.n_R, natom)
@@ -354,7 +358,7 @@ def stupid_centering_fc3_v3(tensor3, check_for_symmetries = True, Far = 1):
             tensor3.tensor = fc3[0:ntrip]
             tensor3.x_r_vector2 = np.rint(np.dot(r_vector2[0:ntrip,:], irprim), dtype=float).T
             tensor3.x_r_vector3 = np.rint(np.dot(r_vector3[0:ntrip,:], irprim), dtype=float).T
-            write_fc3(tensor3)
+            #write_fc3(tensor3)
             if(check_for_symmetries):
                 permutation = thermal_conductivity.third_order_cond_centering.check_permutation_symmetry(tensor3.tensor, tensor3.r_vector2.T, tensor3.r_vector3.T, tensor3.n_R, natom)
                 if(not permutation):
@@ -396,257 +400,6 @@ def write_fc3(tensor3):
                                 for k1 in range(3):
                                     outfile.write(3*' ' + format(tensor3.tensor[i,3*iat+j,3*jat+j1,3*kat+k1], '.8f'))
                             outfile.write('\n')
-
-def stupid_centering_fc3_v2(tensor3, Far = 1):
-
-    rprim = tensor3.unitcell_structure.unit_cell.copy()
-    irprim = np.linalg.inv(rprim)
-    rsup = tensor3.supercell_structure.unit_cell.copy()
-    irsup = np.linalg.inv(rsup)
-    #print(rprim)
-    positions = tensor3.unitcell_structure.coords.copy()
-    xpos = np.dot(positions, np.linalg.inv(rprim))
-    natom = len(xpos)
-    #print(xpos)
-    symbols = tensor3.unitcell_structure.atoms
-    unique_symbols = np.unique(symbols)
-    unique_numbers = np.arange(len(unique_symbols), dtype=int) + 1
-    numbers = np.zeros(len(symbols))
-    for iat in range(len(symbols)):
-        for jat in range(len(unique_symbols)):
-            if(symbols[iat] == unique_symbols[jat]):
-                numbers[iat] = unique_numbers[jat]
-    #print(numbers)
-    cell = (rprim, xpos, numbers)
-
-    if(tensor3.n_R == tensor3.n_sup**2):
-        print('Not previously centered. Stupid centering!')
-        got = [False for x in range(tensor3.n_R)]
-        pairs = []
-        # Check if it satisfies the permutation symmetry
-        for i in range(tensor3.n_R):
-            rvec2 = tensor3.r_vector2[:,i].copy()
-            rvec3 = tensor3.r_vector3[:,i].copy()
-            if(np.linalg.norm(rvec2) < 1.0e-5 and np.linalg.norm(rvec3) < 1.0e-5):
-                pairs.append([i,i])
-                got[i] = True
-            else:
-                if(not got[i]):
-                    for j in range(i, tensor3.n_R):
-                        if(not got[j]):
-                            rvec21 = tensor3.r_vector2[:,j].copy()
-                            rvec31 = tensor3.r_vector3[:,j].copy()
-                            if(np.linalg.norm(rvec21 - rvec3) < 1.0e-5 and np.linalg.norm(rvec31 - rvec2) < 1.0e-5):
-                                pairs.append([i,j])
-                                got[i] = True
-                                got[j] = True
-        if(not np.all(got)):
-            for i in range(tensor3.n_R):
-                if(not got[i]):
-                    print(tensor3.x_r_vector2[:,i])
-                    print(tensor3.x_r_vector3[:,i])
-                    print('')
-        else:
-            print('Found all pairs')
-            for ipair in range(len(pairs)):
-                if(pairs[ipair][0] != pairs[ipair][1]):
-                    ip = pairs[ipair][0]
-                    jp = pairs[ipair][1]
-                    for i in range(3*natom):
-                        for iat in range(natom):
-                            for jat in range(natom):
-                                if(np.any(np.abs(tensor3.tensor[ip,i,3*iat:3*(iat + 1), 3*jat:3*(jat+1)] - tensor3.tensor[jp,i,3*jat:3*(jat + 1), 3*iat:3*(iat+1)].T) > 1.0e-10*np.amax(np.abs(tensor3.tensor)))):
-                                    print('Permutation symmetry failed original!')
-                                    print(tensor3.tensor[ip,i])
-                                    print(tensor3.tensor[jp,i])
-                                    print(tensor3.tensor[ip,i] - tensor3.tensor[jp,i].T)
-                                    print(np.abs(tensor3.tensor[ip,i] - tensor3.tensor[jp,i].T) > 1.0e-8*np.amax(np.abs(tensor3.tensor[ip,i])))
-                                    raise RuntimeError('Get me out!', np.amax(np.abs(tensor3.tensor[ip,i])))
-                                    break
-        new_r_vector2 = [[] for iat in range(natom**3)]
-        new_r_vector3 = [[] for iat in range(natom**3)]
-        new_tensor = [[] for iat in range(natom**3)]
-        multiplicity = [[] for iat in range(natom**3)]
-        print('Finished check. Centering...')
-        for ir in range(tensor3.n_R):
-            rvec2 = tensor3.r_vector2[:,ir].copy()
-            rvec3 = tensor3.r_vector3[:,ir].copy()
-            xvec2 = np.dot(rvec2, irsup)
-            xvec3 = np.dot(rvec3, irsup)
-            size2 = np.linalg.norm(rvec2)
-            size20 = np.linalg.norm(rvec2)
-            size3 = np.linalg.norm(rvec3)
-            size30 = np.linalg.norm(rvec3)
-            size = np.linalg.norm(rvec3 - rvec2) + size2 + size3
-            rvec2_new = [[] for iat in range(natom**3)]
-            rvec3_new = [[] for iat in range(natom**3)]
-            # Find the shortest pair vectors in the mirror supercells -Far <= x <= Far
-            for iat in range(natom):
-                for jat in range(natom):
-                    for kat in range(natom):
-                        index = kat + jat*natom + natom**2*iat
-                        rvec2 = tensor3.r_vector2[:,ir].copy() + positions[jat] - positions[iat]
-                        rvec3 = tensor3.r_vector3[:,ir].copy() + positions[kat] - positions[iat]
-                        xvec2 = np.dot(rvec2, irsup)
-                        xvec3 = np.dot(rvec3, irsup)
-                        size2 = np.linalg.norm(rvec2)
-                        size20 = np.linalg.norm(rvec2)
-                        size3 = np.linalg.norm(rvec3)
-                        size30 = np.linalg.norm(rvec3)
-                        for i in range(-Far,Far + 1):
-                            for j in range(-Far, Far + 1):
-                                for k in range(-Far, Far +1):
-                                    xvec21 = xvec2 + np.array([i,j,k])
-                                    rvec21 = np.dot(xvec21, rsup)
-                                    size21 = np.linalg.norm(rvec21)
-                                    if(size21 <= size2 + 1.0e-6):
-                                        if(abs(size21 - size2) < 1.0e-6):
-                                            uc_vec = tensor3.r_vector2[:,ir].copy()
-                                            uc_vec = np.dot(uc_vec, irsup)
-                                            uc_vec += np.array([i,j,k])
-                                            uc_vec = np.dot(uc_vec, rsup)
-                                            rvec2_new[index].append(uc_vec)
-                                        else:
-                                            uc_vec = tensor3.r_vector2[:,ir].copy()
-                                            uc_vec = np.dot(uc_vec, irsup)
-                                            uc_vec += np.array([i,j,k])
-                                            uc_vec = np.dot(uc_vec, rsup)
-                                            rvec2_new[index] = []
-                                            rvec2_new[index].append(uc_vec)
-                                        size2 = size21
-                                        rvec2 = rvec21.copy()
-                                    xvec31 = xvec3 + np.array([i,j,k])
-                                    rvec31 = np.dot(xvec31, rsup)
-                                    size31 = np.linalg.norm(rvec31)
-                                    if(size31 <= size3 + 1.0e-6):
-                                        if(abs(size31 - size3) < 1.0e-6):
-                                            uc_vec = tensor3.r_vector3[:,ir].copy()
-                                            uc_vec = np.dot(uc_vec, irsup)
-                                            uc_vec += np.array([i,j,k])
-                                            uc_vec = np.dot(uc_vec, rsup)
-                                            rvec3_new[index].append(uc_vec)
-                                        else:
-                                            uc_vec = tensor3.r_vector3[:,ir].copy()
-                                            uc_vec = np.dot(uc_vec, irsup)
-                                            uc_vec += np.array([i,j,k])
-                                            uc_vec = np.dot(uc_vec, rsup)
-                                            rvec3_new[index] = []
-                                            rvec3_new[index].append(uc_vec)
-                                        size3 = size31
-                                        rvec3 = rvec31.copy()
-                # For each pair of the shortest pairs construct another entry to tensor3 and scale it with multiplicity
-                        for iuc in range(len(rvec2_new[index])):
-                            for juc in range(len(rvec3_new[index])):
-                                already_there = False
-                                for kuc in range(len(new_r_vector2[index])):
-                                    if(np.linalg.norm(rvec2_new[index][iuc] - new_r_vector2[index][kuc]) < 1.0e-6 and \
-                                            np.linalg.norm(rvec3_new[index][juc] - new_r_vector3[index][kuc]) < 1.0e-6):
-                                        already_there = True
-                                        print('Would double count this one!')
-                                        break
-                                if not already_there:
-                                    new_r_vector2[index].append(rvec2_new[index][iuc])
-                                    new_r_vector3[index].append(rvec3_new[index][juc])
-                                    new_tensor[index].append(tensor3.tensor[ir]/float(len(rvec2_new[index])*len(rvec3_new[index])))
-                                    multiplicity[index].append(len(rvec2_new[index])*len(rvec3_new[index]))
-        n_R = [0 for x in range(natom**3)]
-        for iat in range(natom):
-            for jat in range(natom):
-                for kat in range(natom):
-                    n_R[kat + jat*natom +iat*natom**2] = len(new_r_vector2[kat + jat*natom + natom**2*iat])
-        #Check if all n_R are equal
-        if(n_R.count(n_R[0]) == len(n_R)):
-            index0 = 0
-        else:
-            index0 = np.argsort(n_R)[-1]
-        r_vector2 = new_r_vector2[index0].copy()
-        r_vector3 = new_r_vector3[index0].copy()
-        extra = 0
-        for itrip in range(len(n_R)):
-            if(itrip != index0):
-                for iuc in range(len(new_r_vector2[itrip])):
-                    already_there = False
-                    for juc in range(len(r_vector2)):
-                        if(np.linalg.norm(r_vector2[juc] - new_r_vector2[itrip][iuc]) < 1.0e-6 and \
-                                np.linalg.norm(r_vector3[juc] - new_r_vector3[itrip][iuc]) < 1.0e-6):
-                            already_there = True
-                            break
-                    if(not already_there):
-                        extra += 1
-                        r_vector2.append(new_r_vector2[itrip][iuc])
-                        r_vector3.append(new_r_vector3[itrip][iuc])
-        print('Added ' + str(extra) + ' new triplets!')
-        fc3 = []
-        for iuc in range(len(r_vector2)):
-            fc3.append(np.zeros_like(tensor3.tensor[0]))
-            for iat in range(natom):
-                for jat in range(natom):
-                    for kat in range(natom):
-                        index = kat + jat*natom + iat*natom**2
-                        found = False
-                        for juc in range(len(new_r_vector2[index])):
-                            if(np.linalg.norm(new_r_vector2[index][juc] - r_vector2[iuc]) < 1.0e-6 and \
-                                    np.linalg.norm(new_r_vector3[index][juc] - r_vector3[iuc]) < 1.0e-6):
-                                fc3[iuc][3*iat:3+3*iat,3*jat:3+3*jat,3*kat:3+3*kat] = new_tensor[index][juc][3*iat:3+3*iat,3*jat:3+3*jat,3*kat:3+3*kat]#*multiplicity[index][juc]/multiplicity[index][iuc]
-                                found = True
-                                break
-        print('Final number of triplets: ', len(r_vector2))
-        tensor3.n_R = len(r_vector2)
-        tensor3.r_vector2 = np.array(r_vector2).T
-        tensor3.r_vector3 = np.array(r_vector3).T
-        tensor3.x_r_vector2 = np.zeros_like(tensor3.r_vector2)
-        tensor3.x_r_vector3 = np.zeros_like(tensor3.r_vector3)
-        tensor3.tensor = np.array(fc3)
-        write_fc3(tensor3)
-        got = [False for x in range(tensor3.n_R)]
-        pairs = []
-        for i in range(tensor3.n_R):
-            rvec2 = tensor3.r_vector2[:,i].copy()
-            rvec3 = tensor3.r_vector3[:,i].copy()
-            tensor3.x_r_vector2[:,i] = np.rint(np.dot(tensor3.r_vector2[:,i], irprim), dtype=float)
-            tensor3.x_r_vector3[:,i] = np.rint(np.dot(tensor3.r_vector3[:,i], irprim), dtype=float)
-            # Find pairs to check if this centering broke permutation symmetry
-            if(np.linalg.norm(rvec2 - rvec3) < 1.0e-5):
-                pairs.append([i,i])
-                got[i] = True
-            else:
-                if(not got[i]):
-                    for j in range(i, tensor3.n_R):
-                        if(not got[j]):
-                            rvec21 = tensor3.r_vector2[:,j].copy()
-                            rvec31 = tensor3.r_vector3[:,j].copy()
-                            if(np.linalg.norm(rvec21 - rvec3) < 1.0e-5 and np.linalg.norm(rvec31 - rvec2) < 1.0e-5):
-                                pairs.append([i,j])
-                                got[i] = True
-                                got[j] = True
-        if(not np.all(got)):
-            for i in range(tensor3.n_R):
-                if(not got[i]):
-                    print(tensor3.x_r_vector2[:,i])
-                    print(tensor3.x_r_vector3[:,i])
-                    print('')
-        else:
-            print('Found all pairs')
-            for ipair in range(len(pairs)):
-                if(pairs[ipair][0] != pairs[ipair][1]):
-                    ip = pairs[ipair][0]
-                    jp = pairs[ipair][1]
-                    for i in range(3*natom):
-                        for iat in range(natom):
-                            for jat in range(natom):
-                                if(np.any(np.abs(tensor3.tensor[ip,i,3*iat:3*(iat + 1), 3*jat:3*(jat+1)] - tensor3.tensor[jp,i,3*jat:3*(jat + 1), 3*iat:3*(iat+1)].T) > 1.0e-10*np.amax(np.abs(tensor3.tensor)))):
-                                    print('Permutation symmetry failed!')
-                                    print(tensor3.tensor[ip,i])
-                                    print(tensor3.tensor[jp,i])
-                                    print(tensor3.tensor[ip,i] - tensor3.tensor[jp,i].T)
-                                    print(np.abs(tensor3.tensor[ip,i] - tensor3.tensor[jp,i].T) > 1.0e-8*np.amax(np.abs(tensor3.tensor[ip,i])))
-                                    raise RuntimeError('Get me out!', np.amax(np.abs(tensor3.tensor[ip,i])))
-                                    break
-
-    else:
-        print('Probably already centered! Nothing to do!')
-
-    return tensor3
 
 def find_q_mq_pairs(kpoints):
 
@@ -1186,7 +939,7 @@ class ThermalConductivity:
                         self.sigmas[ikpt][iband] = delta_v*delta_q*self.smearing_scale
                     else:
                         self.sigmas[ikpt][iband] = 0.0
-            min_smear = np.amax(self.sigmas)/10.0 # We can't have scattering zero for modes with 0 group velocity, so we set it to this number!
+            min_smear = np.amax(self.sigmas)/10.0 # We can't have smearing zero for modes with 0 group velocity, so we set it to this number!
             self.sigmas[self.sigmas < min_smear] = min_smear
             if(np.all(self.sigmas == 0.0)):
                 raise RuntimeError('All smearing values are zero!')
@@ -1706,8 +1459,8 @@ class ThermalConductivity:
             self.delta_omega = energies[1] - energies[0]
         ls_key = format(temperature, '.1f')
 
-        if(mode_mixing != 'no'):
-            print('WARNING! mode_mixing approach has been selected. Calculation of kappa in GK will not be possible!')
+        #if(mode_mixing != 'no'):
+        #    print('WARNING! mode_mixing approach has been selected. Calculation of kappa in GK will not be possible!')
 
         if(method == 'python'):
 
@@ -1902,23 +1655,28 @@ class ThermalConductivity:
     def get_lineshapes_along_the_line(self, temperature, ne = 1000, filename = 'spectral_function_along_path', gauss_smearing = True, mode_mixing = 'no', kpoints = None, start_nkpts = 100, smear = None):
 
         """
-        Calculate phonon lineshapes in full Brillouin zone.
+        Calculate phonon lineshapes for specific k-points.
 
         temperature      : temperature to calculate lineshapes on.
         ne               : Number of frequency points for the lineshapes
         gauss_smearing   : are we using Gaussian smearing as approximation for energy conservation
+        mode_mixing      : If true will calculate full phonon spectral function
         kpoints          : the list of kpoints in reduced coordinates to calculate lineshapes for.
                            If not provided generate them using seekpath
-        nkpts            : Number of k points along the path. Will differ from the final number of points!
-        """
+        start_nkpts      : Number of k points along the path. Will differ from the final number of points!
+        smear            : Smearing used for energy conservation.
 
+        """
+        
+        if(not __SEEKPATH and kpoints is None):
+            raise RuntimeError('To automatically generated a line in reciprocal space one need seekpath. First do "pip install seekpath"!')
+        
         start_time = time.time()
 
         tics = []
         distances = []
         segments = []
         if(kpoints is None):
-            import seekpath
             rat = np.dot(self.dyn.structure.coords, np.linalg.inv(self.dyn.structure.unit_cell))
             sym = np.unique(self.dyn.structure.atoms)
             nt = np.zeros(len(rat))
@@ -1962,6 +1720,7 @@ class ThermalConductivity:
         irrqgrid = kpoints.T
         scattering_events = np.zeros(nkpts, dtype=int)
         if(smear is None):
+            # One has to provide smearing. Otherwise we are using the largest smearing from the full Brillouin zone!
             sigmas = np.zeros((nkpts, self.nband))
             sigmas[:,:] = np.amax(self.sigmas)
         else:
@@ -2000,7 +1759,6 @@ class ThermalConductivity:
         if(self.cp_mode == 'classical'):
             classical = True
 
-        print(psutil.virtual_memory().percent)
         if(mode_mixing == 'mode_mixing'):
             curr_ls = thermal_conductivity.get_lf.calculate_lineshapes_mode_mixing(irrqgrid, scattering_grids, weights, scattering_events,\
                     self.fc2.tensor, self.fc2.r_vector2, self.fc3.tensor, self.fc3.r_vector2, self.fc3.r_vector3, \
