+++ requirements.txt
@@ -3,3 +3,4 @@ numpy
 scipy
 ase
 spglib
+h5py
